backend/replicate-async-status/requirements.txt
psycopg2-binary==2.9.9
replicate==0.34.1